refactor: Faq domain 정리
